Add support for solo mining. Additional improvements for CPU sieve. If no NXS address is specified in the miner.conf file or as a parameter then the miner will work in solo mode.

+++ oacc/AccSieve.h
@@ -11,7 +11,7 @@
 #ifdef __cplusplus
 extern "C" {
 #endif 
-	void pgisieve(unsigned int * sieve, unsigned int sieveSize, mpz_t zPrimorial, mpz_t zPrimeOrigin, unsigned long long ktuple_origin, unsigned int * primes, unsigned int * inverses, unsigned int nPrimorialEndPrime, unsigned int nPrimeLimit, mpz_t * zFirstSieveElement, unsigned long * candidates);
+	void pgisieve(uint64_t * sieve, unsigned int sieveSize, mpz_t zPrimorial, mpz_t zPrimeOrigin, unsigned long long ktuple_origin, unsigned int * primes, unsigned int * inverses, unsigned int nPrimorialEndPrime, unsigned int nPrimeLimit, mpz_t * zFirstSieveElement, unsigned long * candidates);
 	
 #ifdef __cplusplus
 }
